chore: abstracted utils to reuse in tutorial
